+++ easyGame/MainModule/View/GameScreenView.swift
@@ -80,4 +80,5 @@ class GameScreenView: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
 }
