+++ easyGame/MainModule/View/GamepadView.swift
@@ -116,4 +116,5 @@ class GamepadView: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
 }
