+++ easyGame/MainModule/View/MainViewController.swift
@@ -78,4 +78,5 @@ extension MainViewController {
             gamepadView.heightAnchor.constraint(equalToConstant: sqareSize)
         ])
     }
+    
 }
